[refactor] #109 kakao login sdk 사용 로직  view model -> activity로 이동
